+++ alias.config.js
@@ -5,6 +5,7 @@ const r = (p) => resolve(__dirname, p)
 module.exports.alias = {
     '@': r('./src'),
     components: r('./src/components'),
+    common: r('./src/common'),
     hooks: r('./src/hooks'),
     interfaces: r('./src/interfaces'),
     lib: r('./src/lib')