+++ src/components/Navbar/NavbarBrand.tsx
@@ -12,11 +12,18 @@ export interface NavbarBrandProps extends React.HTMLAttributes<HTMLDivElement> {
 const NavbarBrand = ({ hiddenIn, name, imgSrc, imgHeight = 30, imgWidth, ...props }: NavbarBrandProps) => {
     return (
         <div {...props} className={composeClasses('items-center gap-2', hiddenIn ? `hidden ${hiddenIn}:flex` : 'flex', props.className)}>
-            {imgSrc &&
-                <figure className="block">
-                    <img style={{ height: imgHeight, width: imgWidth }} src={imgSrc} />
-                </figure>}
-            {name}
+            {
+                imgSrc
+                    ? (
+                        <>
+                            {<figure className="block">
+                                <img style={{ height: imgHeight, width: imgWidth }} src={imgSrc} />
+                            </figure>}
+                            {name}
+                        </>
+                    )
+                    : props.children
+            }
         </div>
     )
 }