+++ src/components/Navbar/NavbarItem.tsx
@@ -14,8 +14,8 @@ const colorVariants: { [key: string]: string } = {
     warning: 'yellow'
 }
 
-const getActiveVariants = (navLinkVariant: NavbarContentVariants, activeColor: string) => {
-    switch (navLinkVariant) {
+export const getActiveItemVariants = (navItemVariant: NavbarContentVariants, activeColor: string) => {
+    switch (navItemVariant) {
         case 'underline':
             return `h-16 border-b-3 border-${activeColor}-500`
         case 'highlight':
@@ -30,7 +30,7 @@ const NavbarItem = ({ children, isActive, ...props }: NavbarItemProps) => {
 
     const classes = composeClasses(
         'flex content-center items-center flex-nowrap',
-        isActive && getActiveVariants(navbarContentContext.variant, colorVariants[navbarContentContext.activeColor]),
+        isActive && getActiveItemVariants(navbarContentContext.variant, colorVariants[navbarContentContext.activeColor]),
         isActive ? 'font-bold' : 'text-base',
         props.className
     )