+++ src/stories/Navbar.stories.tsx
@@ -51,8 +51,7 @@ const Template: ComponentStory<typeof Navbar> = (args) => (
                 </Navbar.CollapseContent>
             </Navbar.Collapse>
 
-            <Navbar.Brand hiddenIn="sm" imgSrc='http://lendd3r-frontend-develop.s3-website-us-west-1.amazonaws.com/assets/lendder.53f51ab2.png' />
-
+            <Navbar.Brand hiddenIn="sm" imgSrc="http://lendd3r-frontend-develop.s3-website-us-west-1.amazonaws.com/assets/lendder.53f51ab2.png" />
         </Navbar.Group>
 
         <Navbar.Content hiddenIn="sm" activeColor="primary" variant="underline">
@@ -97,7 +96,7 @@ const Template: ComponentStory<typeof Navbar> = (args) => (
                             <UserCircleIcon height={20} width={20} className="text-gray-400" />
                             Mi perfil
                         </div>
-                        <span onClick={() => { }} className="flex rounded-lg gap-2 px-4 py-2 text-sm text-gray-600 hover:bg-gray-100 cursor-pointer">
+                        <span onClick={() => {}} className="flex rounded-lg gap-2 px-4 py-2 text-sm text-gray-600 hover:bg-gray-100 cursor-pointer">
                             <LockClosedIcon className="text-gray-400" height={20} width={20} />
                             Cerrar sesión
                         </span>